refactor: 팩토리 클래스 패키지 이동
