+++ src/main/java/controller/AttendanceController.java
@@ -4,6 +4,7 @@ import domain.AbsentPolicy;
 import domain.AttendanceDateTime;
 import domain.AttendanceSheet;
 import domain.AttendanceSheets;
+import domain.AttendanceSheetsFactory;
 import domain.AttendanceState;
 import util.FileReaderUtil;
 import view.InputView;
